+++ package.json
@@ -13,15 +13,15 @@
   "engines": {
     "node": ">=18"
   },
-  "main": "./dist/index.cjs",
-  "module": "./dist/index.cjs.mjs",
+  "main": "./dist/index.js",
+  "module": "./dist/index.mjs",
   "types": "./dist/index.d.ts",
-  "type": "module",
+  "type": "commonjs",
   "exports": {
     ".": {
       "types": "./dist/index.d.ts",
-      "require": "./dist/index.cjs",
-      "import": "./dist/index.cjs.mjs"
+      "require": "./dist/index.js",
+      "import": "./dist/index.mjs"
     },
     "./package.json": "./package.json"
   },
@@ -36,8 +36,8 @@
     "test:it:browserstack": "npm run build-test-resources && TS_NODE_PROJECT=./tsconfig.json wdio test-config/wdio.browserstack.conf.js",
     "test:it:docker:chrome": "npm run build-test-resources && TS_NODE_PROJECT=./tsconfig.json wdio test-config/wdio.docker.chrome.conf.js",
     "test:it:local:chrome": "npm run build-test-resources && TS_NODE_PROJECT=./tsconfig.json wdio test-config/wdio.local.chrome.conf.js",
-    "build": "rollup -c rollup/dist.config.js",
-    "build-test-resources": "rollup -c rollup/test-resources.config.js",
+    "build": "rollup -c rollup/dist.config.mjs",
+    "build-test-resources": "rollup -c rollup/test-resources.config.mjs",
     "lint": "eslint 'src/**/*.ts' 'test/**/*.ts'",
     "fix-js": "eslint --fix 'src/**/*.ts' 'test/**/*.ts' || exit 0",
     "release:ci:prepatch": "release-it patch --preRelease=alpha-$(git rev-parse --short HEAD) --ci",
@@ -59,6 +59,8 @@
     "@babel/preset-env": "^7.22.5",
     "@babel/runtime": "^7.22.5",
     "@rollup/plugin-babel": "^6.0.3",
+    "@rollup/plugin-commonjs": "^25.0.4",
+    "@rollup/plugin-node-resolve": "^15.2.1",
     "@rollup/plugin-replace": "^5.0.2",
     "@rollup/plugin-strip": "^3.0.2",
     "@rollup/plugin-terser": "^0.4.3",
