--- rollup/test-resources.config.js
@@ -1,23 +1,0 @@
-import replace from '@rollup/plugin-replace'
-import fs from 'fs'
-import path from 'path'
-import config from './dist.config.js'
-
-const packageJson = JSON.parse(fs.readFileSync(path.resolve('package.json'), { encoding: 'utf-8' }))
-
-const OUTPUT_DIR = './test-resources'
-
-export default {
-  input: './test/it/helpers/preambled.ts',
-  output: {
-    file: `${OUTPUT_DIR}/bundle.iife.js`,
-    format: 'iife'
-  },
-  plugins: [
-    ...config[0].plugins,
-    replace({
-      preventAssignment: true,
-      LC_VERSION: JSON.stringify(`${packageJson.versionPrefix}${packageJson.version}`)
-    })
-  ]
-}
